requirements.txt
APScheduler==3.6.3
atomicwrites==1.4.1
attrs==23.1.0
cachetools==4.2.2
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==2.0.12
colorama==0.4.6
flake8==3.9.2
flake8-docstrings==1.6.0
idna==3.6
iniconfig==2.0.0
isort==5.13.2
mccabe==0.6.1
mypy==1.8.0
mypy-extensions==1.0.0
packaging==23.2
pluggy==1.3.0
py==1.11.0
pycodestyle==2.7.0
pydocstyle==6.3.0
pyflakes==2.3.1
pytest==6.2.5
pytest-timeout==2.1.0
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2023.3.post1
requests==2.26.0
six==1.16.0
snowballstemmer==2.2.0
toml==0.10.2
tomli==2.0.1
tornado==6.4
types-backports.ssl-match-hostname==3.7.4.4
types-cachetools==5.3.0.7
types-colorama==0.4.15.20240106
types-pycurl==7.45.2.20240106
types-pytz==2023.3.1.1
types-requests==2.31.0.20240106
types-ujson==5.9.0.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==1.26.18
